webscraping/requirement.txt
beautifulsoup4==4.12.3
Flask==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
